Upgrade to rusqlite 0.38 with native WASM support
rusqlite 0.38.0 (Dec 2024) adds wasm32-unknown-unknown support via
sqlite-wasm-rs, eliminating the need for separate native/WASM code paths.

Changes:
- Upgrade rusqlite from 0.32 to 0.38
- Remove conditional "native" feature from geocoder-core
- Enable rusqlite for all builds (native + WASM)
- Implement actual SQLite query in worker's ShardLoader
- Fix u64 -> i64 for COUNT query (rusqlite 0.38 breaking change)

The worker can now fetch SQLite shards from R2 and query them using
the same geocoder-core Database implementation used for native builds.

+++ crates/geocoder-core/src/database.rs
@@ -135,10 +135,10 @@ impl Database {
 
     /// Get the number of records in the divisions table.
     pub fn count(&self) -> Result<u64> {
-        let count: u64 = self
+        let count: i64 = self
             .conn
             .query_row("SELECT COUNT(*) FROM divisions", [], |row| row.get(0))?;
-        Ok(count)
+        Ok(count as u64)
     }
 
     /// Get metadata value by key.
